+++ services/job-service/src/main/kotlin/com/jobsearch/jobservice/requests/JobRequest.kt
@@ -4,7 +4,7 @@ data class JobRequest(
     val title: String,
     val description: String,
     val location: String,
-    val salary: Double?,
+    val salary: Int?,
     val requiredSkills: String,
     val requiredExperience: String
 )
